começando implementação de Guerreiros

+++ POO/trabalhoPOO/trabalho1POO/src/guerreiros/atlantes/Satiro.java
@@ -12,10 +12,10 @@ import arena.Guerreiro;
  */
 public class Satiro extends GuerreiroAtlante{
     
-    public Satiro(String nome, int idade, int peso) {
+    public Satiro(String nome, int idade, double peso) {
         super(nome, idade, peso);
     }
-
+    
     
     
 }
